refactor(peer-sync): split peer_sync_engine.rs into module directory
Extracts the 2554-line monolith into four focused sub-modules while
keeping mod.rs as the orchestrator. Wire format, transport key
derivation, and session protocol sequence are bit-for-bit identical.

- protocol.rs   — Msg enum, SyncMeta/SyncBookRow/SyncSignalRow, port formula
- crypto.rs     — derive_sync_key_static, derive_sync_key_ecdh, AES-GCM helpers
- connection.rs — TCP frame I/O, plaintext and encrypted message send/recv
- conflict.rs   — all DB helpers (manifests, upserts, LWW merge)
- mod.rs        — SyncEngineState, server/client orchestration, Tauri commands

All 6 unit tests pass (cargo test peer_sync_engine).
