+++ src-tauri/src/commands/peer_sync_engine/mod.rs
@@ -45,13 +45,19 @@
 //!
 //! Frame format: [4-byte big-endian length][12-byte nonce][AES-256-GCM ciphertext]
 
-use aes_gcm::{aead::Aead, Aes256Gcm, KeyInit, Nonce};
+mod conflict;
+mod connection;
+mod crypto;
+mod protocol;
+
+use conflict::*;
+use connection::*;
+use crypto::*;
+use protocol::*;
+pub use protocol::sync_port_for_device;
+
 use chrono::Utc;
-use rand::RngCore;
-use rusqlite::Connection;
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
-use std::io::{Read, Write};
+use serde::Serialize;
 use std::net::{TcpListener, TcpStream, ToSocketAddrs};
 use std::sync::{
     atomic::{AtomicBool, Ordering},
@@ -65,143 +71,6 @@ use x25519_dalek::{EphemeralSecret, PublicKey as X25519PublicKey};
 use crate::commands::peer_identity::get_or_create_device_identity;
 use crate::commands::peer_pairing::{load_trusted_devices, remove_trusted_device};
 use crate::db::{Database, JournalEntryRow};
-
-// ── Port formula ──────────────────────────────────────────────────────────────
-
-/// Derive the sync port from a device_id.
-/// Port = 44000 + (first 4 hex chars as u16) % 1000  →  range 44000–44999.
-pub fn sync_port_for_device(device_id: &str) -> u16 {
-    let hex = &device_id[..device_id.len().min(4)];
-    let offset = u16::from_str_radix(hex, 16).unwrap_or(0) % 1000;
-    44000 + offset
-}
-
-// ── Protocol messages ─────────────────────────────────────────────────────────
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(tag = "t", rename_all = "snake_case")]
-enum Msg {
-    Hello {
-        did: String,
-        /// Ephemeral X25519 public key (hex) for forward-secret session key.
-        /// Absent in pre-v2 peers — falls back to static key derivation.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        eph_pub: Option<String>,
-    },
-    Ok {
-        name: String,
-        /// Ephemeral X25519 public key (hex), present when server supports v2.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        eph_pub: Option<String>,
-    },
-    /// Sent plaintext by the server when the connecting device is not in its
-    /// trusted list. The client should auto-revoke the server in response.
-    NotTrusted {
-        server_device_id: String,
-    },
-    Manifest {
-        entries: Vec<SyncMeta>,
-        books: Vec<SyncMeta>,
-        /// Signals use created_at as the version field (signals are immutable).
-        signals: Vec<SyncMeta>,
-        /// Settings manifest: only whitelisted keys; updated_at for LWW.
-        settings: Vec<SyncMeta>,
-    },
-    // ── Entry phase ──
-    Entry {
-        row: JournalEntryRow,
-    },
-    Done {
-        sent: usize,
-    },
-    DoneAck {
-        recv: usize,
-    },
-    // ── Books phase ──
-    Book {
-        row: SyncBookRow,
-    },
-    BooksDone {
-        sent: usize,
-    },
-    BooksAck {
-        recv: usize,
-    },
-    // ── Signals phase ──
-    Signal {
-        row: SyncSignalRow,
-    },
-    SignalsDone {
-        sent: usize,
-    },
-    SignalsAck {
-        recv: usize,
-    },
-    // ── Settings phase ──
-    Setting {
-        key: String,
-        value: String,
-        updated_at: String,
-    },
-    SettingsDone {
-        sent: usize,
-    },
-    SettingsAck {
-        recv: usize,
-    },
-    // ── Full restore protocol ──
-    /// Client → Server: request the server's complete DB as a binary stream.
-    RestoreRequest,
-    /// Server → Client: metadata envelope preceding each binary chunk frame.
-    RestoreChunk {
-        seq: u64,
-        total_chunks: u64,
-        offset: u64,
-        total_bytes: u64,
-    },
-    /// Server → Client: all chunks have been sent.
-    RestoreEnd {
-        total_bytes: u64,
-        chunks: u64,
-    },
-    Err {
-        msg: String,
-    },
-}
-
-/// Generic manifest item used for all data types.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct SyncMeta {
-    id: String,
-    updated_at: String,
-}
-
-/// Book row transmitted over the wire during sync.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct SyncBookRow {
-    id: String,
-    name: String,
-    emoji: String,
-    color: String,
-    sort_order: i32,
-    description: Option<String>,
-    settings: Option<String>,
-    created_at: String,
-    updated_at: String,
-}
-
-/// Signal row transmitted over the wire during sync.
-/// Signals are immutable after creation — no updated_at.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct SyncSignalRow {
-    id: String,
-    timestamp: String,
-    #[serde(rename = "type")]
-    signal_type: String,
-    source: String,
-    payload: String,
-    created_at: String,
-}
 
 // ── Managed state ─────────────────────────────────────────────────────────────
 
@@ -229,620 +98,6 @@ impl SyncEngineState {
 
 unsafe impl Send for SyncEngineState {}
 unsafe impl Sync for SyncEngineState {}
-
-// ── Crypto helpers ────────────────────────────────────────────────────────────
-
-/// Legacy: static key from both Ed25519 public keys (no forward secrecy).
-/// Used as fallback when the remote peer does not advertise an eph_pub.
-fn derive_sync_key_static(pub_a: &str, pub_b: &str) -> [u8; 32] {
-    let mut keys = [pub_a, pub_b];
-    keys.sort_unstable();
-    let mut h = Sha256::new();
-    h.update(b"moodhaven-sync-v1:");
-    h.update(keys[0].as_bytes());
-    h.update(keys[1].as_bytes());
-    h.finalize().into()
-}
-
-/// v2: ephemeral X25519 ECDH + static identity binding → forward-secret session key.
-/// session_key = SHA-256("moodhaven-sync-v2:" || X25519_shared || sorted(static_a, static_b))
-fn derive_sync_key_ecdh(
-    my_eph_secret: EphemeralSecret,
-    peer_eph_pub_hex: &str,
-    my_static_pub: &str,
-    peer_static_pub: &str,
-) -> Result<[u8; 32], String> {
-    let peer_bytes =
-        hex::decode(peer_eph_pub_hex).map_err(|e| format!("bad peer eph_pub hex: {e}"))?;
-    let peer_arr: [u8; 32] = peer_bytes
-        .try_into()
-        .map_err(|_| "peer eph_pub must be 32 bytes".to_string())?;
-    let peer_pub = X25519PublicKey::from(peer_arr);
-    let shared = my_eph_secret.diffie_hellman(&peer_pub);
-
-    let mut static_keys = [my_static_pub, peer_static_pub];
-    static_keys.sort_unstable();
-    let mut h = Sha256::new();
-    h.update(b"moodhaven-sync-v2:");
-    h.update(shared.as_bytes());
-    h.update(static_keys[0].as_bytes());
-    h.update(static_keys[1].as_bytes());
-    Ok(h.finalize().into())
-}
-
-/// Encrypt plaintext → [12-byte nonce][ciphertext]
-fn encrypt_payload(key: &[u8; 32], plaintext: &[u8]) -> Result<Vec<u8>, String> {
-    let cipher = Aes256Gcm::new_from_slice(key).map_err(|e| format!("AES init: {e}"))?;
-    let mut nonce_bytes = [0u8; 12];
-    rand::rngs::OsRng.fill_bytes(&mut nonce_bytes);
-    let nonce = Nonce::from_slice(&nonce_bytes);
-    let ct = cipher
-        .encrypt(nonce, plaintext)
-        .map_err(|e| format!("Encrypt: {e}"))?;
-    let mut out = nonce_bytes.to_vec();
-    out.extend(ct);
-    Ok(out)
-}
-
-/// Decrypt [12-byte nonce][ciphertext] → plaintext
-fn decrypt_payload(key: &[u8; 32], data: &[u8]) -> Result<Vec<u8>, String> {
-    if data.len() < 12 {
-        return Err("Frame too short to decrypt".into());
-    }
-    let cipher = Aes256Gcm::new_from_slice(key).map_err(|e| format!("AES init: {e}"))?;
-    let (nonce_bytes, ct) = data.split_at(12);
-    let nonce = Nonce::from_slice(nonce_bytes);
-    cipher
-        .decrypt(nonce, ct)
-        .map_err(|_| "Decryption failed (wrong key or tampered data)".to_string())
-}
-
-// ── Frame I/O ─────────────────────────────────────────────────────────────────
-
-/// Write a length-prefixed frame: [4-byte big-endian length][payload]
-fn write_frame(stream: &mut TcpStream, payload: &[u8]) -> Result<(), String> {
-    let len = payload.len() as u32;
-    stream
-        .write_all(&len.to_be_bytes())
-        .map_err(|e| format!("write frame length: {e}"))?;
-    stream
-        .write_all(payload)
-        .map_err(|e| format!("write frame payload: {e}"))?;
-    Ok(())
-}
-
-/// Read exactly the next length-prefixed frame bytes.
-/// The limit is generous (256 MB) to accommodate large binary restore chunks
-/// while still guarding against malformed frames.
-fn read_frame_bytes(stream: &mut TcpStream) -> Result<Vec<u8>, String> {
-    let mut len_buf = [0u8; 4];
-    stream
-        .read_exact(&mut len_buf)
-        .map_err(|e| format!("read frame length: {e}"))?;
-    let len = u32::from_be_bytes(len_buf) as usize;
-    if len > 256 * 1024 * 1024 {
-        return Err(format!("Frame too large: {len} bytes"));
-    }
-    let mut buf = vec![0u8; len];
-    stream
-        .read_exact(&mut buf)
-        .map_err(|e| format!("read frame payload: {e}"))?;
-    Ok(buf)
-}
-
-/// Write a raw binary frame (no encryption — only used for DB chunk data
-/// which is already sent after an encrypted envelope confirming chunk metadata).
-fn write_binary_frame(stream: &mut TcpStream, data: &[u8]) -> Result<(), String> {
-    let len = data.len() as u32;
-    stream
-        .write_all(&len.to_be_bytes())
-        .map_err(|e| format!("write binary frame length: {e}"))?;
-    stream
-        .write_all(data)
-        .map_err(|e| format!("write binary frame data: {e}"))?;
-    Ok(())
-}
-
-/// Read a raw binary frame (no decryption — pair with write_binary_frame).
-fn read_binary_frame(stream: &mut TcpStream) -> Result<Vec<u8>, String> {
-    read_frame_bytes(stream)
-}
-
-// ── Message I/O ───────────────────────────────────────────────────────────────
-
-/// Send a plaintext JSON message (used for HELLO/OK before key exchange).
-fn write_msg(stream: &mut TcpStream, msg: &Msg) -> Result<(), String> {
-    let json = serde_json::to_vec(msg).map_err(|e| format!("serialize msg: {e}"))?;
-    write_frame(stream, &json)
-}
-
-/// Send an AES-GCM encrypted JSON message.
-fn write_msg_enc(stream: &mut TcpStream, key: &[u8; 32], msg: &Msg) -> Result<(), String> {
-    let json = serde_json::to_vec(msg).map_err(|e| format!("serialize msg: {e}"))?;
-    let payload = encrypt_payload(key, &json)?;
-    write_frame(stream, &payload)
-}
-
-/// Read and parse a plaintext JSON message.
-fn read_msg(stream: &mut TcpStream) -> Result<Msg, String> {
-    let bytes = read_frame_bytes(stream)?;
-    serde_json::from_slice(&bytes).map_err(|e| format!("parse msg: {e}"))
-}
-
-/// Read and decrypt an AES-GCM encrypted JSON message.
-fn read_msg_enc(stream: &mut TcpStream, key: &[u8; 32]) -> Result<Msg, String> {
-    let bytes = read_frame_bytes(stream)?;
-    let plain = decrypt_payload(key, &bytes)?;
-    serde_json::from_slice(&plain).map_err(|e| format!("parse decrypted msg: {e}"))
-}
-
-// ── DB helpers ────────────────────────────────────────────────────────────────
-
-fn db_get_entries_manifest(conn: &Connection) -> Result<Vec<SyncMeta>, String> {
-    let mut stmt = conn
-        .prepare("SELECT id, updated_at FROM journal_entries ORDER BY updated_at DESC")
-        .map_err(|e| format!("prepare entries manifest: {e}"))?;
-    let rows = stmt
-        .query_map([], |r| {
-            Ok(SyncMeta {
-                id: r.get(0)?,
-                updated_at: r.get(1)?,
-            })
-        })
-        .map_err(|e| format!("query entries manifest: {e}"))?;
-    rows.collect::<Result<Vec<_>, _>>()
-        .map_err(|e| format!("collect entries manifest: {e}"))
-}
-
-fn db_get_books_manifest(conn: &Connection) -> Result<Vec<SyncMeta>, String> {
-    let mut stmt = conn
-        .prepare("SELECT id, COALESCE(updated_at, created_at) FROM books ORDER BY id")
-        .map_err(|e| format!("prepare books manifest: {e}"))?;
-    let rows = stmt
-        .query_map([], |r| {
-            Ok(SyncMeta {
-                id: r.get(0)?,
-                updated_at: r.get(1)?,
-            })
-        })
-        .map_err(|e| format!("query books manifest: {e}"))?;
-    rows.collect::<Result<Vec<_>, _>>()
-        .map_err(|e| format!("collect books manifest: {e}"))
-}
-
-fn db_get_signals_manifest(conn: &Connection) -> Result<Vec<SyncMeta>, String> {
-    let mut stmt = conn
-        .prepare("SELECT id, created_at FROM signals ORDER BY created_at DESC")
-        .map_err(|e| format!("prepare signals manifest: {e}"))?;
-    let rows = stmt
-        .query_map([], |r| {
-            Ok(SyncMeta {
-                id: r.get(0)?,
-                updated_at: r.get(1)?,
-            })
-        })
-        .map_err(|e| format!("query signals manifest: {e}"))?;
-    rows.collect::<Result<Vec<_>, _>>()
-        .map_err(|e| format!("collect signals manifest: {e}"))
-}
-
-/// Returns manifest items for whitelisted settings keys that exist in the DB.
-fn db_get_settings_manifest(conn: &Connection) -> Result<Vec<SyncMeta>, String> {
-    let result: rusqlite::Result<(String, String)> = conn.query_row(
-        "SELECT key, updated_at FROM settings WHERE key = 'app_settings'",
-        [],
-        |r| Ok((r.get(0)?, r.get(1)?)),
-    );
-    match result {
-        Ok((key, updated_at)) => Ok(vec![SyncMeta {
-            id: key,
-            updated_at,
-        }]),
-        Err(rusqlite::Error::QueryReturnedNoRows) => Ok(vec![]),
-        Err(e) => Err(format!("query settings manifest: {e}")),
-    }
-}
-
-fn db_get_books_full(conn: &Connection, ids: &[String]) -> Result<Vec<SyncBookRow>, String> {
-    if ids.is_empty() {
-        return Ok(vec![]);
-    }
-    let placeholders = ids.iter().map(|_| "?").collect::<Vec<_>>().join(", ");
-    let sql = format!(
-        "SELECT id, name, emoji, color, sort_order, description, settings, \
-                created_at, COALESCE(updated_at, created_at) \
-         FROM books WHERE id IN ({placeholders})"
-    );
-    conn.prepare(&sql)
-        .map_err(|e| format!("prepare books full: {e}"))?
-        .query_map(rusqlite::params_from_iter(ids.iter()), |r| {
-            Ok(SyncBookRow {
-                id: r.get(0)?,
-                name: r.get(1)?,
-                emoji: r.get(2)?,
-                color: r.get(3)?,
-                sort_order: r.get(4)?,
-                description: r.get(5)?,
-                settings: r.get(6)?,
-                created_at: r.get(7)?,
-                updated_at: r.get(8)?,
-            })
-        })
-        .map_err(|e| format!("query books full: {e}"))?
-        .collect::<rusqlite::Result<Vec<_>>>()
-        .map_err(|e| format!("collect books full: {e}"))
-}
-
-fn db_upsert_book(conn: &Connection, row: &SyncBookRow) -> Result<bool, String> {
-    let existing: Option<String> = conn
-        .query_row(
-            "SELECT COALESCE(updated_at, created_at) FROM books WHERE id = ?1",
-            rusqlite::params![row.id],
-            |r| r.get(0),
-        )
-        .ok();
-
-    match existing {
-        None => {
-            conn.execute(
-                "INSERT INTO books \
-                 (id, name, emoji, color, sort_order, description, settings, created_at, updated_at) \
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
-                rusqlite::params![
-                    row.id, row.name, row.emoji, row.color, row.sort_order,
-                    row.description, row.settings, row.created_at, row.updated_at
-                ],
-            )
-            .map_err(|e| format!("insert book: {e}"))?;
-            Ok(true)
-        }
-        Some(ref local) if row.updated_at.as_str() > local.as_str() => {
-            conn.execute(
-                "UPDATE books \
-                 SET name = ?2, emoji = ?3, color = ?4, sort_order = ?5, \
-                     description = ?6, settings = ?7, updated_at = ?8 \
-                 WHERE id = ?1",
-                rusqlite::params![
-                    row.id,
-                    row.name,
-                    row.emoji,
-                    row.color,
-                    row.sort_order,
-                    row.description,
-                    row.settings,
-                    row.updated_at
-                ],
-            )
-            .map_err(|e| format!("update book: {e}"))?;
-            Ok(true)
-        }
-        _ => Ok(false), // local is same age or newer — skip
-    }
-}
-
-fn db_get_signals_full(conn: &Connection, ids: &[String]) -> Result<Vec<SyncSignalRow>, String> {
-    if ids.is_empty() {
-        return Ok(vec![]);
-    }
-    let placeholders = ids.iter().map(|_| "?").collect::<Vec<_>>().join(", ");
-    let sql = format!(
-        "SELECT id, timestamp, type, source, payload, created_at \
-         FROM signals WHERE id IN ({placeholders})"
-    );
-    conn.prepare(&sql)
-        .map_err(|e| format!("prepare signals full: {e}"))?
-        .query_map(rusqlite::params_from_iter(ids.iter()), |r| {
-            Ok(SyncSignalRow {
-                id: r.get(0)?,
-                timestamp: r.get(1)?,
-                signal_type: r.get(2)?,
-                source: r.get(3)?,
-                payload: r.get(4)?,
-                created_at: r.get(5)?,
-            })
-        })
-        .map_err(|e| format!("query signals full: {e}"))?
-        .collect::<rusqlite::Result<Vec<_>>>()
-        .map_err(|e| format!("collect signals full: {e}"))
-}
-
-/// Signals are immutable — INSERT OR IGNORE; returns true if a new row was inserted.
-fn db_insert_signal_if_new(conn: &Connection, row: &SyncSignalRow) -> Result<bool, String> {
-    let changes = conn
-        .execute(
-            "INSERT OR IGNORE INTO signals (id, timestamp, type, source, payload, created_at) \
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-            rusqlite::params![
-                row.id,
-                row.timestamp,
-                row.signal_type,
-                row.source,
-                row.payload,
-                row.created_at
-            ],
-        )
-        .map_err(|e| format!("insert signal: {e}"))?;
-    Ok(changes > 0)
-}
-
-fn db_get_setting_for_sync(
-    conn: &Connection,
-    key: &str,
-) -> Result<Option<(String, String)>, String> {
-    let result: rusqlite::Result<(String, String)> = conn.query_row(
-        "SELECT value, updated_at FROM settings WHERE key = ?1",
-        rusqlite::params![key],
-        |r| Ok((r.get(0)?, r.get(1)?)),
-    );
-    match result {
-        Ok(pair) => Ok(Some(pair)),
-        Err(rusqlite::Error::QueryReturnedNoRows) => Ok(None),
-        Err(e) => Err(format!("query setting: {e}")),
-    }
-}
-
-/// Merge whitelisted fields from `remote` JSON into `local` JSON.
-/// Non-whitelisted fields (credentials, device-specific prefs) are never overwritten.
-fn merge_settings_json(local_json: &str, remote_json: &str) -> Result<String, String> {
-    let mut local: serde_json::Value =
-        serde_json::from_str(local_json).map_err(|e| format!("parse local settings: {e}"))?;
-    let remote: serde_json::Value =
-        serde_json::from_str(remote_json).map_err(|e| format!("parse remote settings: {e}"))?;
-
-    // Take these top-level sections entirely from remote
-    for section in &["journal", "reminders"] {
-        if let Some(v) = remote.get(*section) {
-            local[section] = v.clone();
-        }
-    }
-    // Take only ai.features and ai.consent from remote (not ai.openai / ai.localAI / ai.enabled)
-    if let Some(remote_ai) = remote.get("ai") {
-        if let Some(local_ai) = local.get_mut("ai") {
-            for sub in &["features", "consent"] {
-                if let Some(v) = remote_ai.get(*sub) {
-                    local_ai[sub] = v.clone();
-                }
-            }
-        }
-    }
-    // Take specific appearance fields from remote (not theme — that's per-device)
-    if let Some(remote_app) = remote.get("appearance") {
-        if let Some(local_app) = local.get_mut("appearance") {
-            for field in &["compactMode", "animationsEnabled"] {
-                if let Some(v) = remote_app.get(*field) {
-                    local_app[field] = v.clone();
-                }
-            }
-        }
-    }
-
-    serde_json::to_string(&local).map_err(|e| format!("serialize merged settings: {e}"))
-}
-
-/// Upsert a setting received from a peer, applying whitelist merge for app_settings.
-/// Returns true if the local DB was changed.
-fn db_upsert_setting(
-    conn: &Connection,
-    key: &str,
-    remote_value: &str,
-    remote_updated_at: &str,
-) -> Result<bool, String> {
-    let local = db_get_setting_for_sync(conn, key)?;
-    let new_value = match &local {
-        None => remote_value.to_string(),
-        Some((local_value, local_updated_at)) => {
-            if remote_updated_at <= local_updated_at.as_str() {
-                return Ok(false); // local is same age or newer
-            }
-            if key == "app_settings" {
-                merge_settings_json(local_value, remote_value)?
-            } else {
-                remote_value.to_string()
-            }
-        }
-    };
-
-    conn.execute(
-        "INSERT OR REPLACE INTO settings (key, value, updated_at) VALUES (?1, ?2, ?3)",
-        rusqlite::params![key, new_value, remote_updated_at],
-    )
-    .map_err(|e| format!("upsert setting: {e}"))?;
-    Ok(true)
-}
-
-fn db_get_entries_full(conn: &Connection, ids: &[String]) -> Result<Vec<JournalEntryRow>, String> {
-    if ids.is_empty() {
-        return Ok(vec![]);
-    }
-    let placeholders = ids.iter().map(|_| "?").collect::<Vec<_>>().join(", ");
-    let sql = format!(
-        "SELECT je.id, je.encrypted_content, je.mood, je.privacy_mode, \
-                je.location_weather, je.book_id, je.pinned, \
-                je.created_at, je.updated_at, \
-                je.sealed_until, je.capsule_type, je.linked_original_id, je.unsealed_at, \
-                COALESCE(GROUP_CONCAT(t.name, ','), '') AS tags \
-         FROM journal_entries je \
-         LEFT JOIN entry_tags et ON et.entry_id = je.id \
-         LEFT JOIN tags t ON t.id = et.tag_id \
-         WHERE je.id IN ({placeholders}) \
-         GROUP BY je.id"
-    );
-    conn.prepare(&sql)
-        .map_err(|e| format!("prepare entries full: {e}"))?
-        .query_map(rusqlite::params_from_iter(ids.iter()), |r| {
-            let ec_json: String = r.get(1)?;
-            let tags_str: Option<String> = r.get(13)?;
-            let tags = crate::db::parse_tags(tags_str);
-            let ec: crate::db::EncryptedContent = serde_json::from_str(&ec_json).map_err(|e| {
-                rusqlite::Error::FromSqlConversionFailure(
-                    1,
-                    rusqlite::types::Type::Text,
-                    Box::new(std::io::Error::new(
-                        std::io::ErrorKind::InvalidData,
-                        e.to_string(),
-                    )),
-                )
-            })?;
-            Ok(JournalEntryRow {
-                id: r.get(0)?,
-                encrypted_content: Some(ec),
-                mood: r.get(2)?,
-                privacy_mode: r.get(3)?,
-                location_weather: r.get(4)?,
-                book_id: r
-                    .get::<_, Option<String>>(5)?
-                    .unwrap_or_else(|| "default".to_string()),
-                pinned: r.get::<_, i32>(6)? != 0,
-                created_at: r.get(7)?,
-                updated_at: r.get(8)?,
-                sealed_until: r.get(9)?,
-                capsule_type: r.get(10)?,
-                linked_original_id: r.get(11)?,
-                unsealed_at: r.get(12)?,
-                tags,
-                status: None,
-            })
-        })
-        .map_err(|e| format!("query entries full: {e}"))?
-        .collect::<rusqlite::Result<Vec<_>>>()
-        .map_err(|e| format!("collect entries full: {e}"))
-}
-
-fn db_upsert_tags(conn: &Connection, entry_id: &str, tags: &[String]) -> Result<(), String> {
-    if tags.is_empty() {
-        return Ok(());
-    }
-    for tag_name in tags {
-        let tag_name = tag_name.trim();
-        if tag_name.is_empty() {
-            continue;
-        }
-        // Ensure tag exists
-        conn.execute(
-            "INSERT OR IGNORE INTO tags (id, name) VALUES (lower(hex(randomblob(8))), ?1)",
-            rusqlite::params![tag_name],
-        )
-        .map_err(|e| format!("insert tag: {e}"))?;
-
-        // Get tag id
-        let tag_id: String = conn
-            .query_row(
-                "SELECT id FROM tags WHERE name = ?1",
-                rusqlite::params![tag_name],
-                |r| r.get(0),
-            )
-            .map_err(|e| format!("get tag id: {e}"))?;
-
-        // Link tag to entry
-        conn.execute(
-            "INSERT OR IGNORE INTO entry_tags (entry_id, tag_id) VALUES (?1, ?2)",
-            rusqlite::params![entry_id, tag_id],
-        )
-        .map_err(|e| format!("link tag: {e}"))?;
-    }
-    Ok(())
-}
-
-fn db_upsert_entry(conn: &Connection, row: &JournalEntryRow) -> Result<bool, String> {
-    let ec = row
-        .encrypted_content
-        .as_ref()
-        .ok_or("encrypted_content is None — cannot sync sealed entry")?;
-    let ec_json = serde_json::to_string(ec).map_err(|e| format!("serialize ec: {e}"))?;
-
-    // Check existing updated_at
-    let existing: Option<String> = conn
-        .query_row(
-            "SELECT updated_at FROM journal_entries WHERE id = ?1",
-            rusqlite::params![row.id],
-            |r| r.get(0),
-        )
-        .ok();
-
-    match existing {
-        None => {
-            // INSERT — trigger only fires on UPDATE, so updated_at is preserved as-is
-            conn.execute(
-                "INSERT INTO journal_entries \
-                 (id, encrypted_content, mood, privacy_mode, location_weather, \
-                  book_id, pinned, created_at, updated_at, \
-                  sealed_until, capsule_type, linked_original_id, unsealed_at) \
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13)",
-                rusqlite::params![
-                    row.id,
-                    ec_json,
-                    row.mood,
-                    row.privacy_mode,
-                    row.location_weather,
-                    row.book_id,
-                    row.pinned as i32,
-                    row.created_at,
-                    row.updated_at,
-                    row.sealed_until,
-                    row.capsule_type,
-                    row.linked_original_id,
-                    row.unsealed_at,
-                ],
-            )
-            .map_err(|e| format!("insert entry: {e}"))?;
-            db_upsert_tags(conn, &row.id, &row.tags)?;
-            Ok(true)
-        }
-        Some(ref local) if row.updated_at.as_str() > local.as_str() => {
-            // UPDATE — set updated_at explicitly so the trigger (WHEN NEW.updated_at = OLD.updated_at) doesn't fire
-            conn.execute(
-                "UPDATE journal_entries \
-                 SET encrypted_content = ?2, mood = ?3, privacy_mode = ?4, \
-                     location_weather = ?5, book_id = ?6, pinned = ?7, updated_at = ?8, \
-                     sealed_until = ?9, capsule_type = ?10, \
-                     linked_original_id = ?11, unsealed_at = ?12 \
-                 WHERE id = ?1",
-                rusqlite::params![
-                    row.id,
-                    ec_json,
-                    row.mood,
-                    row.privacy_mode,
-                    row.location_weather,
-                    row.book_id,
-                    row.pinned as i32,
-                    row.updated_at,
-                    row.sealed_until,
-                    row.capsule_type,
-                    row.linked_original_id,
-                    row.unsealed_at,
-                ],
-            )
-            .map_err(|e| format!("update entry: {e}"))?;
-            db_upsert_tags(conn, &row.id, &row.tags)?;
-            Ok(true)
-        }
-        _ => Ok(false), // local is same age or newer — skip
-    }
-}
-
-#[allow(dead_code)]
-fn db_get_peer_sync_at(conn: &Connection, peer_id: &str) -> Result<Option<String>, String> {
-    let result: rusqlite::Result<String> = conn.query_row(
-        "SELECT last_sync_at FROM peer_sync_state WHERE peer_device_id = ?1",
-        rusqlite::params![peer_id],
-        |r| r.get(0),
-    );
-    match result {
-        Ok(val) => Ok(Some(val)),
-        Err(rusqlite::Error::QueryReturnedNoRows) => Ok(None),
-        Err(e) => Err(format!("query peer_sync_at: {e}")),
-    }
-}
-
-fn db_set_peer_sync_at(conn: &Connection, peer_id: &str, at: &str) -> Result<(), String> {
-    conn.execute(
-        "INSERT INTO peer_sync_state (peer_device_id, last_sync_at) VALUES (?1, ?2)
-         ON CONFLICT(peer_device_id) DO UPDATE SET last_sync_at = excluded.last_sync_at",
-        rusqlite::params![peer_id, at],
-    )
-    .map_err(|e| format!("set peer sync at: {e}"))?;
-    Ok(())
-}
 
 // ── Full-restore server handler ───────────────────────────────────────────────
 
@@ -2385,6 +1640,7 @@ pub fn peer_apply_and_restart(app: AppHandle) -> Result<(), String> {
 mod tests {
     use super::*;
     use crate::db::{EncryptedContent, JournalEntryRow};
+    use rusqlite::Connection;
 
     // ── Helpers ───────────────────────────────────────────────────────────────
